+++ packages/frontend/src/App.tsx
@@ -5,9 +5,9 @@ import '@blueprintjs/datetime2/lib/css/blueprint-datetime2.css';
 import '@blueprintjs/popover2/lib/css/blueprint-popover2.css';
 import '@blueprintjs/select/lib/css/blueprint-select.css';
 import '@blueprintjs/table/lib/css/table.css';
-import moment from 'moment';
 import React from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
+import { ReactQueryDevtools } from 'react-query/devtools';
 import {
     BrowserRouter as Router,
     Redirect,
@@ -48,11 +48,6 @@ import { AppProvider } from './providers/AppProvider';
 import { DashboardProvider } from './providers/DashboardProvider';
 import { TrackingProvider, TrackPage } from './providers/TrackingProvider';
 import { PageName } from './types/Events';
-
-// Keep original date time values on filters instead of converting dates into UTC when using JSON.stringify on API requests
-Date.prototype.toJSON = function () {
-    return moment(this).format('YYYY-MM-DDTHH:mm:ss');
-};
 
 const queryClient = new QueryClient({
     defaultOptions: {
@@ -315,6 +310,7 @@ const App = () => (
                     </TrackingProvider>
                 </AppProvider>
             </HotkeysProvider>
+            <ReactQueryDevtools initialIsOpen={false} />
         </QueryClientProvider>
     </>
 );